backend/boq_engine_v1/requirements.txt
pandas>=2.2
pyyaml>=6.0.1
# IFC
ifcopenshell>=0.7
# DXF
ezdxf>=1.1.4
# Export
openpyxl>=3.1
